Question.id change to int, add some incorrect logic.

+++ src/main/java/com/epam/db/service/QuestionService.java
@@ -5,5 +5,5 @@ import com.epam.telegram.entity.Question;
 import java.util.List;
 
 public interface QuestionService {
-    Question getQuestionWithoutRepetition(List<Long> ids);
+    Question getQuestionWithoutRepetition(List<Integer> ids);
 }
